mv fillDefinitionVariables -> useFillVariables
