+++ src/contexts/Help/defaults.ts
@@ -1,12 +1,7 @@
 // Copyright 2022 @paritytech/polkadot-staking-dashboard authors & contributors
 // SPDX-License-Identifier: Apache-2.0
 
-import { HelpContextInterface, HelpDefinition } from './types';
-
-export const definition: HelpDefinition = {
-  title: '',
-  description: [],
-};
+import { HelpContextInterface } from './types';
 
 export const defaultHelpContext: HelpContextInterface = {
   // eslint-disable-next-line @typescript-eslint/no-unused-vars
@@ -16,8 +11,6 @@ export const defaultHelpContext: HelpContextInterface = {
   setStatus: (s) => {},
   // eslint-disable-next-line @typescript-eslint/no-unused-vars
   setDefinition: (d) => {},
-  // eslint-disable-next-line @typescript-eslint/no-unused-vars
-  fillDefinitionVariables: (d) => definition,
   status: 0,
   definition: null,
 };
