+++ src/contexts/Help/types.ts
@@ -32,7 +32,6 @@ export interface HelpContextInterface {
   closeHelp: () => void;
   setStatus: (s: number) => void;
   setDefinition: (d: MaybeString) => void;
-  fillDefinitionVariables: (d: HelpDefinition) => HelpDefinition;
   status: number;
   definition: MaybeString;
 }
