+++ src/library/Help/index.tsx
@@ -13,21 +13,16 @@ import {
   HelpItemRaw,
 } from 'contexts/Help/types';
 import { useAnimation } from 'framer-motion';
+import useFillVariables from 'library/Hooks/useFillVariables';
 import { useEffect } from 'react';
 import Definition from './Items/Definition';
 import External from './Items/External';
 import { ContentWrapper, HeightWrapper, Wrapper } from './Wrappers';
 
 export const Help = () => {
-  const {
-    setStatus,
-    status,
-    fillDefinitionVariables,
-    definition,
-    closeHelp,
-    setDefinition,
-  } = useHelp();
+  const { setStatus, status, definition, closeHelp, setDefinition } = useHelp();
   const controls = useAnimation();
+  const { fillVariables } = useFillVariables();
 
   const onFadeIn = async () => {
     await controls.start('visible');
@@ -91,9 +86,7 @@ export const Help = () => {
     : null;
 
   // fill placeholder variables
-  activeDefinition = activeDefinition
-    ? fillDefinitionVariables(activeDefinition)
-    : null;
+  activeDefinition = activeDefinition ? fillVariables(activeDefinition) : null;
 
   // filter active definition
   definitions = definitions.filter(
@@ -154,7 +147,7 @@ export const Help = () => {
                   Definitions
                 </h3>
                 {definitions.map((item: HelpDefinition, index: number) => {
-                  item = fillDefinitionVariables(item);
+                  item = fillVariables(item);
                   return (
                     <Definition
                       key={`def_${index}`}
