+++ src/library/Hooks/useFillVariables/types.ts
@@ -1,0 +1,7 @@
+// Copyright 2022 @paritytech/polkadot-staking-dashboard authors & contributors
+// SPDX-License-Identifier: Apache-2.0
+
+export interface FillVariableItem {
+  title: string;
+  description: string[];
+}
